split up driver.go
