+++ internal/store/driver.go
@@ -1,7 +1,6 @@
 package store
 
 import (
-	"cmp"
 	"context"
 	"encoding/json"
 	"fmt"
@@ -11,23 +10,19 @@ import (
 	"sync"
 
 	"github.com/go-logr/logr"
-	"golang.org/x/exp/slices"
 	corev1 "k8s.io/api/core/v1"
 	netv1 "k8s.io/api/networking/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/utils/ptr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	gatewayv1 "sigs.k8s.io/gateway-api/apis/v1"
 
-	common "github.com/ngrok/ngrok-operator/api/common/v1alpha1"
 	ingressv1alpha1 "github.com/ngrok/ngrok-operator/api/ingress/v1alpha1"
 	ngrokv1alpha1 "github.com/ngrok/ngrok-operator/api/ngrok/v1alpha1"
 
 	"github.com/ngrok/ngrok-operator/internal/annotations"
 	"github.com/ngrok/ngrok-operator/internal/errors"
 	"github.com/ngrok/ngrok-operator/internal/util"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 const defaultClusterDomain = "svc.cluster.local"
@@ -477,168 +472,6 @@ func (d *Driver) Sync(ctx context.Context, c client.Client) error {
 	return nil
 }
 
-func (d *Driver) SyncEdges(ctx context.Context, c client.Client) error {
-	if !d.syncAllowConcurrent {
-		if proceed, wait := d.syncStart(true); proceed {
-			defer d.syncDone()
-		} else {
-			return wait(ctx)
-		}
-	}
-
-	d.log.Info("syncing edges state!!")
-	_, desiredIngressDomains, desiredGatewayDomainMap := d.calculateDomains()
-
-	desiredEdges := d.calculateHTTPSEdges(&desiredIngressDomains, desiredGatewayDomainMap)
-	currEdges := &ingressv1alpha1.HTTPSEdgeList{}
-	if err := c.List(ctx, currEdges, client.MatchingLabels{
-		labelControllerNamespace: d.managerName.Namespace,
-		labelControllerName:      d.managerName.Name,
-	}); err != nil {
-		d.log.Error(err, "error listing edges")
-		return err
-	}
-
-	if err := d.applyHTTPSEdges(ctx, c, desiredEdges, currEdges.Items); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (d *Driver) applyDomains(ctx context.Context, c client.Client, desiredDomains, currentDomains []ingressv1alpha1.Domain) error {
-	for _, desiredDomain := range desiredDomains {
-		found := false
-		for _, currDomain := range currentDomains {
-			if desiredDomain.Name == currDomain.Name && desiredDomain.Namespace == currDomain.Namespace {
-				// It matches so lets update it if anything is different
-				if !reflect.DeepEqual(desiredDomain.Spec, currDomain.Spec) {
-					currDomain.Spec = desiredDomain.Spec
-					if err := c.Update(ctx, &currDomain); err != nil {
-						d.log.Error(err, "error updating domain", "domain", desiredDomain)
-						return err
-					}
-				}
-				found = true
-				break
-			}
-		}
-		if !found {
-			if err := c.Create(ctx, &desiredDomain); err != nil {
-				d.log.Error(err, "error creating domain", "domain", desiredDomain)
-				return err
-			}
-		}
-	}
-
-	// Don't delete domains to prevent accidentally de-registering them and making people re-do DNS
-
-	return nil
-}
-
-func (d *Driver) applyHTTPSEdges(ctx context.Context, c client.Client, desiredEdges map[string]ingressv1alpha1.HTTPSEdge, currentEdges []ingressv1alpha1.HTTPSEdge) error {
-	// update or delete edge we don't need anymore
-	for _, currEdge := range currentEdges {
-		hostports := currEdge.Spec.Hostports
-
-		// If one of the controller-owned edges has more than one hostport, log an error and skip it
-		// because we can't determine what to do with it.
-		if len(hostports) != 1 {
-			d.log.Error(nil, "Existing owned edge has more than 1 hostport", "edge", currEdge, "hostports", hostports)
-			continue
-		}
-
-		// ngrok only supports https on port 443 and all domains are on port 443
-		// so we can safely trim the port from the hostport to get the domain
-		domain := strings.TrimSuffix(hostports[0], ":443")
-
-		if desiredEdge, ok := desiredEdges[domain]; ok {
-			needsUpdate := false
-
-			if !reflect.DeepEqual(desiredEdge.Spec, currEdge.Spec) {
-				currEdge.Spec = desiredEdge.Spec
-				needsUpdate = true
-			}
-
-			if needsUpdate {
-				if err := c.Update(ctx, &currEdge); err != nil {
-					d.log.Error(err, "error updating edge", "desiredEdge", desiredEdge, "currEdge", currEdge)
-					return err
-				}
-			}
-
-			// matched and updated the edge, no longer desired
-			delete(desiredEdges, domain)
-		} else {
-			if err := c.Delete(ctx, &currEdge); client.IgnoreNotFound(err) != nil {
-				d.log.Error(err, "error deleting edge", "edge", currEdge)
-				return err
-			}
-		}
-	}
-
-	// the set of desired edges now only contains new edges, create them
-	for _, edge := range desiredEdges {
-		if err := c.Create(ctx, &edge); err != nil {
-			d.log.Error(err, "error creating edge", "edge", edge)
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (d *Driver) applyTunnels(ctx context.Context, c client.Client, desiredTunnels map[tunnelKey]ingressv1alpha1.Tunnel, currentTunnels []ingressv1alpha1.Tunnel) error {
-	// update or delete tunnels we don't need anymore
-	for _, currTunnel := range currentTunnels {
-		// extract tunnel key
-		tkey := d.tunnelKeyFromTunnel(currTunnel)
-
-		// check if new state still needs this tunnel
-		if desiredTunnel, ok := desiredTunnels[tkey]; ok {
-			needsUpdate := false
-
-			// compare/update owner references
-			if !slices.Equal(desiredTunnel.OwnerReferences, currTunnel.OwnerReferences) {
-				needsUpdate = true
-				currTunnel.OwnerReferences = desiredTunnel.OwnerReferences
-			}
-
-			// compare/update desired tunnel spec
-			if !reflect.DeepEqual(desiredTunnel.Spec, currTunnel.Spec) {
-				needsUpdate = true
-				currTunnel.Spec = desiredTunnel.Spec
-			}
-
-			if needsUpdate {
-				if err := c.Update(ctx, &currTunnel); err != nil {
-					d.log.Error(err, "error updating tunnel", "tunnel", desiredTunnel)
-					return err
-				}
-			}
-
-			// matched and updated the tunnel, no longer desired
-			delete(desiredTunnels, tkey)
-		} else {
-			// no longer needed, delete it
-			if err := c.Delete(ctx, &currTunnel); client.IgnoreNotFound(err) != nil {
-				d.log.Error(err, "error deleting tunnel", "tunnel", currTunnel)
-				return err
-			}
-		}
-	}
-
-	// the set of desired tunnels now only contains new tunnels, create them
-	for _, tunnel := range desiredTunnels {
-		if err := c.Create(ctx, &tunnel); err != nil {
-			d.log.Error(err, "error creating tunnel", "tunnel", tunnel)
-			return err
-		}
-	}
-
-	return nil
-}
-
 func (d *Driver) updateIngressStatuses(ctx context.Context, c client.Client) error {
 	ingresses := d.store.ListNgrokIngressesV1()
 	for _, ingress := range ingresses {
@@ -652,86 +485,6 @@ func (d *Driver) updateIngressStatuses(ctx context.Context, c client.Client) err
 		}
 	}
 	return nil
-}
-
-func (d *Driver) calculateDomains() ([]ingressv1alpha1.Domain, []ingressv1alpha1.Domain, map[string]ingressv1alpha1.Domain) {
-	var domains, ingressDomains []ingressv1alpha1.Domain
-	ingressDomainMap := d.calculateDomainsFromIngress()
-
-	ingressDomains = make([]ingressv1alpha1.Domain, 0, len(ingressDomainMap))
-	for _, domain := range ingressDomainMap {
-		ingressDomains = append(ingressDomains, domain)
-		domains = append(domains, domain)
-	}
-
-	var gatewayDomainMap map[string]ingressv1alpha1.Domain
-	if d.gatewayEnabled {
-		gatewayDomainMap = d.calculateDomainsFromGateway(ingressDomainMap)
-		for _, domain := range gatewayDomainMap {
-			domains = append(domains, domain)
-		}
-	}
-
-	return domains, ingressDomains, gatewayDomainMap
-}
-
-func (d *Driver) calculateDomainsFromIngress() map[string]ingressv1alpha1.Domain {
-	domainMap := make(map[string]ingressv1alpha1.Domain)
-
-	ingresses := d.store.ListNgrokIngressesV1()
-	for _, ingress := range ingresses {
-		for _, rule := range ingress.Spec.Rules {
-			if rule.Host == "" {
-				continue
-			}
-
-			domain := ingressv1alpha1.Domain{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      ingressv1alpha1.HyphenatedDomainNameFromURL(rule.Host),
-					Namespace: ingress.Namespace,
-				},
-				Spec: ingressv1alpha1.DomainSpec{
-					Domain: rule.Host,
-				},
-			}
-			domain.Spec.Metadata = d.ingressNgrokMetadata
-			domainMap[rule.Host] = domain
-		}
-	}
-
-	return domainMap
-}
-
-func (d *Driver) calculateDomainsFromGateway(ingressDomains map[string]ingressv1alpha1.Domain) map[string]ingressv1alpha1.Domain {
-	domainMap := make(map[string]ingressv1alpha1.Domain)
-
-	gateways := d.store.ListGateways()
-	for _, gw := range gateways {
-		for _, listener := range gw.Spec.Listeners {
-			if listener.Hostname == nil {
-				continue
-			}
-			domainName := string(*listener.Hostname)
-			if _, hasVal := ingressDomains[domainName]; hasVal {
-				// TODO update gateway status
-				// also add error to error page
-				continue
-			}
-			domain := ingressv1alpha1.Domain{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      ingressv1alpha1.HyphenatedDomainNameFromURL(domainName),
-					Namespace: gw.Namespace,
-				},
-				Spec: ingressv1alpha1.DomainSpec{
-					Domain: domainName,
-				},
-			}
-			domain.Spec.Metadata = d.gatewayNgrokMetadata
-			domainMap[domainName] = domain
-		}
-	}
-
-	return domainMap
 }
 
 // Given an ingress, it will resolve any ngrok modulesets defined on the ingress to the
@@ -770,196 +523,6 @@ func (d *Driver) getNgrokTrafficPolicyForIngress(ing *netv1.Ingress) (*ngrokv1al
 	return d.store.GetNgrokTrafficPolicyV1(policy, ing.Namespace)
 }
 
-func (d *Driver) calculateHTTPSEdges(ingressDomains *[]ingressv1alpha1.Domain, gatewayDomainMap map[string]ingressv1alpha1.Domain) map[string]ingressv1alpha1.HTTPSEdge {
-	edgeMap := make(map[string]ingressv1alpha1.HTTPSEdge, len(*ingressDomains))
-	for _, domain := range *ingressDomains {
-		edge := ingressv1alpha1.HTTPSEdge{
-			ObjectMeta: metav1.ObjectMeta{
-				GenerateName: domain.Name + "-",
-				Namespace:    domain.Namespace,
-				Labels:       d.edgeLabels(),
-			},
-			Spec: ingressv1alpha1.HTTPSEdgeSpec{
-				Hostports: []string{domain.Spec.Domain + ":443"},
-			},
-		}
-		edge.Spec.Metadata = d.ingressNgrokMetadata
-		edgeMap[domain.Spec.Domain] = edge
-	}
-	d.calculateHTTPSEdgesFromIngress(edgeMap)
-
-	if d.gatewayEnabled {
-		gatewayEdgeMap := make(map[string]ingressv1alpha1.HTTPSEdge)
-		httproutes := d.store.ListHTTPRoutes()
-		gateways := d.store.ListGateways()
-		for _, gtw := range gateways {
-			gatewayDomains := make(map[string]string)
-			for _, listener := range gtw.Spec.Listeners {
-				if listener.Hostname == nil {
-					continue
-				}
-				if listener.Protocol != gatewayv1.HTTPSProtocolType || int(listener.Port) != 443 {
-					continue
-				}
-				if _, hasDomain := gatewayDomainMap[string(*listener.Hostname)]; !hasDomain {
-					continue
-				}
-				gatewayDomains[string(*listener.Hostname)] = string(*listener.Hostname)
-			}
-			if len(gatewayDomains) == 0 {
-				d.log.Info("no usable domains in gateway, may be missing https listener", "gateway", gtw.Name)
-				continue
-			}
-			for _, httproute := range httproutes {
-				var routeDomains []string
-				for _, parent := range httproute.Spec.ParentRefs {
-					if string(parent.Name) != gtw.Name {
-						continue
-					}
-					var domainOverlap []string
-					for _, hostname := range httproute.Spec.Hostnames {
-						domain := string(hostname)
-						if _, hasDomain := gatewayDomains[domain]; hasDomain {
-							domainOverlap = append(domainOverlap, domain)
-						}
-					}
-					if len(domainOverlap) == 0 {
-						// no hostnames overlap with gateway
-						continue
-					}
-					routeDomains = append(routeDomains, domainOverlap...)
-				}
-				if len(routeDomains) == 0 {
-					// no usable domains in route
-					continue
-				}
-				var hostPorts []string
-
-				for _, domain := range routeDomains {
-					hostPorts = append(hostPorts, domain+":443")
-				}
-				edge := ingressv1alpha1.HTTPSEdge{
-					ObjectMeta: metav1.ObjectMeta{
-						GenerateName: httproute.Name + "-",
-						Namespace:    httproute.Namespace,
-						Labels:       d.edgeLabels(),
-					},
-					Spec: ingressv1alpha1.HTTPSEdgeSpec{
-						Hostports: hostPorts,
-					},
-				}
-				edge.Spec.Metadata = d.gatewayNgrokMetadata
-				gatewayEdgeMap[routeDomains[0]] = edge
-
-			}
-		}
-		d.calculateHTTPSEdgesFromGateway(gatewayEdgeMap)
-
-		// merge edge maps
-		for k, v := range gatewayEdgeMap {
-			edgeMap[k] = v
-		}
-	}
-
-	return edgeMap
-}
-
-func (d *Driver) calculateHTTPSEdgesFromIngress(edgeMap map[string]ingressv1alpha1.HTTPSEdge) {
-	ingresses := d.store.ListNgrokIngressesV1()
-	for _, ingress := range ingresses {
-		modSet, err := d.getNgrokModuleSetForIngress(ingress)
-		if err != nil {
-			d.log.Error(err, "error getting ngrok moduleset for ingress", "ingress", ingress)
-			continue
-		}
-
-		policyJSON, err := d.getTrafficPolicyJSON(ingress, modSet)
-		if err != nil {
-			d.log.Error(err, "error marshalling JSON Policy for ingress", "ingress", ingress)
-			continue
-		}
-
-		for _, rule := range ingress.Spec.Rules {
-			edge, ok := edgeMap[rule.Host]
-			if !ok {
-				d.log.Error(err, "could not find edge associated with rule", "host", rule.Host)
-				continue
-			}
-
-			if modSet.Modules.TLSTermination != nil && modSet.Modules.TLSTermination.MinVersion != nil {
-				edge.Spec.TLSTermination = &ingressv1alpha1.EndpointTLSTerminationAtEdge{
-					MinVersion: ptr.Deref(modSet.Modules.TLSTermination.MinVersion, ""),
-				}
-			}
-
-			if modSet.Modules.MutualTLS != nil {
-				edge.Spec.MutualTLS = modSet.Modules.MutualTLS
-			}
-
-			// If any rule for an ingress matches, then it applies to this ingress
-			for _, httpIngressPath := range rule.HTTP.Paths {
-				matchType := "path_prefix"
-				if httpIngressPath.PathType != nil {
-					switch *httpIngressPath.PathType {
-					case netv1.PathTypePrefix:
-						matchType = "path_prefix"
-					case netv1.PathTypeExact:
-						matchType = "exact_path"
-					case netv1.PathTypeImplementationSpecific:
-						matchType = "path_prefix" // Path Prefix seems like a sane default for most cases
-					default:
-						d.log.Error(fmt.Errorf("unknown path type"), "unknown path type", "pathType", *httpIngressPath.PathType)
-						continue
-					}
-				}
-
-				// We only support service backends right now. TODO: support resource backends
-				if httpIngressPath.Backend.Service == nil {
-					continue
-				}
-
-				serviceName := httpIngressPath.Backend.Service.Name
-				serviceUID, servicePort, err := d.getEdgeBackend(*httpIngressPath.Backend.Service, ingress.Namespace)
-				if err != nil {
-					d.log.Error(err, "could not find port for service", "namespace", ingress.Namespace, "service", serviceName)
-					continue
-				}
-
-				route := ingressv1alpha1.HTTPSEdgeRouteSpec{
-					Match:     httpIngressPath.Path,
-					MatchType: matchType,
-					Backend: ingressv1alpha1.TunnelGroupBackend{
-						Labels: d.ngrokLabels(ingress.Namespace, serviceUID, serviceName, servicePort),
-					},
-					CircuitBreaker:      modSet.Modules.CircuitBreaker,
-					Compression:         modSet.Modules.Compression,
-					IPRestriction:       modSet.Modules.IPRestriction,
-					Headers:             modSet.Modules.Headers,
-					OAuth:               modSet.Modules.OAuth,
-					Policy:              policyJSON,
-					OIDC:                modSet.Modules.OIDC,
-					SAML:                modSet.Modules.SAML,
-					WebhookVerification: modSet.Modules.WebhookVerification,
-				}
-				route.Metadata = d.ingressNgrokMetadata
-
-				// Loop through existing routes and check if any match the path and match type
-				// If they do, warn about it and continue replacing it
-				for _, existingRoute := range edge.Spec.Routes {
-					if existingRoute.Match == route.Match && existingRoute.MatchType == route.MatchType {
-						d.log.Info("replacing existing route", "route", existingRoute.Match, "newRoute", route.Match)
-						continue
-					}
-				}
-
-				edge.Spec.Routes = append(edge.Spec.Routes, route)
-			}
-
-			edgeMap[rule.Host] = edge
-		}
-	}
-}
-
 // getTrafficPolicyJSON retrieves the traffic policy for an ingress and falls back to the modSet policy if it doesn't exist.
 func (d *Driver) getTrafficPolicyJSON(ingress *netv1.Ingress, modSet *ingressv1alpha1.NgrokModuleSet) (json.RawMessage, error) {
 	var err error
@@ -990,133 +553,6 @@ func (d *Driver) getTrafficPolicyJSON(ingress *netv1.Ingress, modSet *ingressv1a
 	}
 
 	return policyJSON, nil
-}
-
-func (d *Driver) calculateHTTPSEdgesFromGateway(edgeMap map[string]ingressv1alpha1.HTTPSEdge) {
-	gateways := d.store.ListGateways()
-
-	for _, gtw := range gateways {
-		for _, listener := range gtw.Spec.Listeners {
-			if listener.Hostname == nil {
-				continue
-			}
-			allowedRoutes := listener.AllowedRoutes.Kinds
-			if len(allowedRoutes) > 0 {
-				createHttpsedge := false
-				for _, routeKind := range allowedRoutes {
-					if routeKind.Kind == "HTTPRoute" {
-						createHttpsedge = true
-					}
-				}
-				if !createHttpsedge {
-					continue
-				}
-			}
-			domainName := string(*listener.Hostname)
-			edge, ok := edgeMap[domainName]
-			if !ok {
-				continue
-			}
-			// TODO: Calculate routes from httpRoutes
-			// TODO: skip if no backend services
-			httproutes := d.store.ListHTTPRoutes()
-			for _, httproute := range httproutes {
-				for _, parent := range httproute.Spec.ParentRefs {
-					if string(parent.Name) != gtw.Name {
-						// not our gateway so skip
-						continue
-					}
-
-					if listener.AllowedRoutes != nil && listener.AllowedRoutes.Namespaces.From != nil {
-						switch *listener.AllowedRoutes.Namespaces.From {
-						case gatewayv1.NamespacesFromAll:
-						case gatewayv1.NamespacesFromSame:
-							if httproute.Namespace != gtw.Namespace {
-								continue
-							}
-						case gatewayv1.NamespacesFromSelector:
-							if httproute.Namespace != listener.AllowedRoutes.Namespaces.Selector.String() {
-								continue
-							}
-						}
-					}
-
-					// matches our gateway
-					for _, hostname := range httproute.Spec.Hostnames {
-						if string(hostname) != string(*listener.Hostname) {
-							// doesn't match this listener
-							continue
-						}
-						// matches gateway and listener
-						for _, rule := range httproute.Spec.Rules {
-							// TODO: resolve rule.Matches
-							// TODO: resolve rule.Filters
-							// for v0 we will only resolve the first backendRef
-							pathMatch := "/"
-							pathMatchType := "path_prefix"
-							// first match with a path will be accepted as the route's path
-							for _, match := range rule.Matches {
-								if match.Path != nil {
-									pathMatch = *match.Path.Value
-									if *match.Path.Type == gatewayv1.PathMatchExact {
-										pathMatchType = "exact_path"
-									}
-									break
-								}
-							}
-							route := ingressv1alpha1.HTTPSEdgeRouteSpec{
-								Match:     pathMatch,     // change based on the rule.match
-								MatchType: pathMatchType, // change based on rule.Matches
-							}
-
-							// TODO: set with values from rules.Filters + rules.Matches
-							// this HTTPRouteRule comes direct from gateway api yaml, and func returns the policy,
-							// which goes directly into the edge route in ngrok.
-							policy, err := d.createEndpointPolicyForGateway(&rule, httproute.Namespace)
-							if err != nil {
-								d.log.Error(err, "error creating policy from HTTPRouteRule", "rule", rule)
-								continue
-							}
-
-							route.Policy = policy
-
-							for idx, backendref := range rule.BackendRefs {
-								// currently the ingress controller doesn't support weighted backends
-								// so we'll only support one backendref per rule
-								// TODO: remove when tested with multiple backends
-								if idx > 0 {
-									break
-								}
-								// handle backendref
-								refKind := string(*backendref.Kind)
-								if refKind != "Service" {
-									// only support services currently
-									continue
-								}
-
-								refName := string(backendref.Name)
-								serviceUID, servicePort, err := d.getEdgeBackendRef(backendref.BackendRef, httproute.Namespace)
-								if err != nil {
-									d.log.Error(err, "could not find port for service", "namespace", httproute.Namespace, "service", refName)
-									continue
-								}
-
-								route.Backend = ingressv1alpha1.TunnelGroupBackend{
-									Labels: d.ngrokLabels(httproute.Namespace, serviceUID, refName, servicePort),
-								}
-
-							}
-							route.Metadata = d.gatewayNgrokMetadata
-
-							edge.Spec.Routes = append(edge.Spec.Routes, route)
-						}
-					}
-				}
-			}
-
-			edgeMap[domainName] = edge
-		}
-	}
 }
 
 func (d *Driver) createEndpointPolicyForGateway(rule *gatewayv1.HTTPRouteRule, namespace string) (json.RawMessage, error) {
@@ -1573,170 +1009,6 @@ func (d *Driver) handleRequestRedirectFilter(filter *gatewayv1.HTTPRequestRedire
 	return nil
 }
 
-type tunnelKey struct {
-	namespace string
-	service   string
-	port      string
-}
-
-func (d *Driver) tunnelKeyFromTunnel(tunnel ingressv1alpha1.Tunnel) tunnelKey {
-	return tunnelKey{
-		namespace: tunnel.Namespace,
-		service:   tunnel.Labels[labelService],
-		port:      tunnel.Labels[labelPort],
-	}
-}
-
-func (d *Driver) calculateTunnels() map[tunnelKey]ingressv1alpha1.Tunnel {
-	tunnels := map[tunnelKey]ingressv1alpha1.Tunnel{}
-	d.calculateTunnelsFromIngress(tunnels)
-	d.calculateTunnelsFromGateway(tunnels)
-	return tunnels
-}
-
-func (d *Driver) calculateTunnelsFromIngress(tunnels map[tunnelKey]ingressv1alpha1.Tunnel) {
-	for _, ingress := range d.store.ListNgrokIngressesV1() {
-		for _, rule := range ingress.Spec.Rules {
-			for _, path := range rule.HTTP.Paths {
-				// We only support service backends right now.
-				// TODO: support resource backends
-				if path.Backend.Service == nil {
-					continue
-				}
-
-				serviceName := path.Backend.Service.Name
-				serviceUID, servicePort, protocol, appProtocol, err := d.getTunnelBackend(*path.Backend.Service, ingress.Namespace)
-				if err != nil {
-					d.log.Error(err, "could not find port for service", "namespace", ingress.Namespace, "service", serviceName)
-				}
-
-				key := tunnelKey{ingress.Namespace, serviceName, strconv.Itoa(int(servicePort))}
-				tunnel, found := tunnels[key]
-				if !found {
-					targetAddr := fmt.Sprintf("%s.%s.%s:%d", serviceName, key.namespace, d.clusterDomain, servicePort)
-					tunnel = ingressv1alpha1.Tunnel{
-						ObjectMeta: metav1.ObjectMeta{
-							GenerateName:    fmt.Sprintf("%s-%d-", serviceName, servicePort),
-							Namespace:       ingress.Namespace,
-							OwnerReferences: nil, // fill owner references below
-							Labels:          d.tunnelLabels(serviceName, servicePort),
-						},
-						Spec: ingressv1alpha1.TunnelSpec{
-							ForwardsTo: targetAddr,
-							Labels:     d.ngrokLabels(ingress.Namespace, serviceUID, serviceName, servicePort),
-							BackendConfig: &ingressv1alpha1.BackendConfig{
-								Protocol: protocol,
-							},
-						},
-					}
-					switch appProtocol {
-					case "http2":
-						proto := common.ApplicationProtocol_HTTP2
-						tunnel.Spec.AppProtocol = &proto
-					case "http1":
-						proto := common.ApplicationProtocol_HTTP1
-						tunnel.Spec.AppProtocol = &proto
-					}
-				}
-
-				hasIngressReference := false
-				for _, ref := range tunnel.OwnerReferences {
-					if ref.UID == ingress.UID {
-						hasIngressReference = true
-						break
-					}
-				}
-				if !hasIngressReference {
-					tunnel.OwnerReferences = append(tunnel.OwnerReferences, metav1.OwnerReference{
-						APIVersion: ingress.APIVersion,
-						Kind:       ingress.Kind,
-						Name:       ingress.Name,
-						UID:        ingress.UID,
-					})
-					slices.SortStableFunc(tunnel.OwnerReferences, func(i, j metav1.OwnerReference) int {
-						return cmp.Compare(string(i.UID), string(j.UID))
-					})
-				}
-
-				tunnels[key] = tunnel
-			}
-		}
-	}
-}
-
-func (d *Driver) calculateTunnelsFromGateway(tunnels map[tunnelKey]ingressv1alpha1.Tunnel) {
-	httproutes := d.store.ListHTTPRoutes()
-
-	for _, httproute := range httproutes {
-		for _, rule := range httproute.Spec.Rules {
-			for _, backendRef := range rule.BackendRefs {
-				// We only support service backends right now.
-				// TODO: support resource backends
-
-				//if path.Backend.Service == nil {
-				//	continue
-				//}
-
-				serviceName := string(backendRef.Name)
-				serviceUID, servicePort, protocol, appProtocol, err := d.getTunnelBackendFromGateway(backendRef.BackendRef, httproute.Namespace)
-				if err != nil {
-					d.log.Error(err, "could not find port for service", "namespace", httproute.Namespace, "service", serviceName)
-				}
-
-				key := tunnelKey{httproute.Namespace, serviceName, strconv.Itoa(int(servicePort))}
-				tunnel, found := tunnels[key]
-				if !found {
-					targetAddr := fmt.Sprintf("%s.%s.%s:%d", serviceName, key.namespace, d.clusterDomain, servicePort)
-					tunnel = ingressv1alpha1.Tunnel{
-						ObjectMeta: metav1.ObjectMeta{
-							GenerateName:    fmt.Sprintf("%s-%d-", serviceName, servicePort),
-							Namespace:       httproute.Namespace,
-							OwnerReferences: nil, // fill owner references below
-							Labels:          d.tunnelLabels(serviceName, servicePort),
-						},
-						Spec: ingressv1alpha1.TunnelSpec{
-							ForwardsTo: targetAddr,
-							Labels:     d.ngrokLabels(httproute.Namespace, serviceUID, serviceName, servicePort),
-							BackendConfig: &ingressv1alpha1.BackendConfig{
-								Protocol: protocol,
-							},
-						},
-					}
-					switch appProtocol {
-					case "http2":
-						proto := common.ApplicationProtocol_HTTP2
-						tunnel.Spec.AppProtocol = &proto
-					case "http1":
-						proto := common.ApplicationProtocol_HTTP1
-						tunnel.Spec.AppProtocol = &proto
-					}
-				}
-
-				hasReference := false
-				for _, ref := range tunnel.OwnerReferences {
-					if ref.UID == httproute.UID {
-						hasReference = true
-						break
-					}
-				}
-				if !hasReference {
-					tunnel.OwnerReferences = append(tunnel.OwnerReferences, metav1.OwnerReference{
-						APIVersion: httproute.APIVersion,
-						Kind:       httproute.Kind,
-						Name:       httproute.Name,
-						UID:        httproute.UID,
-					})
-					slices.SortStableFunc(tunnel.OwnerReferences, func(i, j metav1.OwnerReference) int {
-						return cmp.Compare(string(i.UID), string(j.UID))
-					})
-				}
-
-				tunnels[key] = tunnel
-			}
-		}
-	}
-}
-
 func (d *Driver) calculateIngressLoadBalancerIPStatus(ing *netv1.Ingress, c client.Reader) []netv1.IngressLoadBalancerIngress {
 	ingressHosts := map[string]bool{}
 	for _, rule := range ing.Spec.Rules {
@@ -1784,27 +1056,6 @@ func (d *Driver) calculateIngressLoadBalancerIPStatus(ing *netv1.Ingress, c clie
 	return status
 }
 
-func (d *Driver) getEdgeBackend(backendSvc netv1.IngressServiceBackend, namespace string) (string, int32, error) {
-	service, servicePort, err := d.findBackendServicePort(backendSvc, namespace)
-	if err != nil {
-		return "", 0, err
-	}
-
-	return string(service.UID), servicePort.Port, nil
-}
-
-func (d *Driver) getEdgeBackendRef(backendRef gatewayv1.BackendRef, namespace string) (string, int32, error) {
-	if backendRef.Namespace != nil && string(*backendRef.Namespace) != namespace {
-		return "", 0, fmt.Errorf("namespace %s not supported", string(*backendRef.Namespace))
-	}
-	service, servicePort, err := d.findBackendRefServicePort(backendRef, namespace)
-	if err != nil {
-		return "", 0, err
-	}
-
-	return string(service.UID), servicePort.Port, nil
-}
-
 func (d *Driver) findBackendRefServicePort(backendRef gatewayv1.BackendRef, namespace string) (*corev1.Service, *corev1.ServicePort, error) {
 	service, err := d.store.GetServiceV1(string(backendRef.Name), namespace)
 	if err != nil {
@@ -1826,44 +1077,6 @@ func (d *Driver) findBackendRefServicesPort(service *corev1.Service, backendRef 
 		}
 	}
 	return nil, fmt.Errorf("could not find matching port for service %s, backend port %v, name %s", service.Name, int32(*backendRef.Port), string(backendRef.Name))
-}
-
-func (d *Driver) getTunnelBackend(backendSvc netv1.IngressServiceBackend, namespace string) (string, int32, string, string, error) {
-	service, servicePort, err := d.findBackendServicePort(backendSvc, namespace)
-	if err != nil {
-		return "", 0, "", "", err
-	}
-
-	protocol, err := d.getPortAnnotatedProtocol(service, servicePort.Name)
-	if err != nil {
-		return "", 0, "", "", err
-	}
-
-	appProtocol, err := d.getPortAppProtocol(service, servicePort)
-	if err != nil {
-		return "", 0, "", "", err
-	}
-
-	return string(service.UID), servicePort.Port, protocol, appProtocol, nil
-}
-
-func (d *Driver) getTunnelBackendFromGateway(backendRef gatewayv1.BackendRef, namespace string) (string, int32, string, string, error) {
-	service, servicePort, err := d.findBackendRefServicePort(backendRef, namespace)
-	if err != nil {
-		return "", 0, "", "", err
-	}
-
-	protocol, err := d.getPortAnnotatedProtocol(service, servicePort.Name)
-	if err != nil {
-		return "", 0, "", "", err
-	}
-
-	appProtocol, err := d.getPortAppProtocol(service, servicePort)
-	if err != nil {
-		return "", 0, "", "", err
-	}
-
-	return string(service.UID), servicePort.Port, protocol, appProtocol, nil
 }
 
 func (d *Driver) findBackendServicePort(backendSvc netv1.IngressServiceBackend, namespace string) (*corev1.Service, *corev1.ServicePort, error) {
@@ -1928,22 +1141,6 @@ func (d *Driver) getPortAppProtocol(service *corev1.Service, port *corev1.Servic
 		return "", nil
 	default:
 		return "", fmt.Errorf("unsupported appProtocol: '%s', must be 'k8s.ngrok.com/http2', 'kubernetes.io/h2c' or ''. From: %s service: %s", proto, service.Namespace, service.Name)
-	}
-}
-
-func (d *Driver) edgeLabels() map[string]string {
-	return map[string]string{
-		labelControllerNamespace: d.managerName.Namespace,
-		labelControllerName:      d.managerName.Name,
-	}
-}
-
-func (d *Driver) tunnelLabels(serviceName string, port int32) map[string]string {
-	return map[string]string{
-		labelControllerNamespace: d.managerName.Namespace,
-		labelControllerName:      d.managerName.Name,
-		labelService:             serviceName,
-		labelPort:                strconv.Itoa(int(port)),
 	}
 }
