+++ internal/store/driver_endpoints.go
@@ -1,0 +1,8 @@
+package store
+
+func (d *Driver) agentEndpointLabels() map[string]string {
+	return map[string]string{
+		labelControllerNamespace: d.managerName.Namespace,
+		labelControllerName:      d.managerName.Name,
+	}
+}
